Intermediary commit for utils.
